Switch license to GPL3
My reading of clause 9 of the GPL2 and the phrase "covered by the GPL" without a version number in src/header is that this is allowed. Prompted by the fact that we link against readline, which is under GPL3.

If I'm wrong, this can all be backed out but then I don't know what to do about readline.